xmlrpcpp: Add support for Nil value type.
XML RPC has an optional Nil data type, which is used by the Python's
xmlrpclib client. Add support for Nil to xmlrpccpp, so that we can
interop with the Python client.

Related changes:
Added a function to explicitly set the type to Nil since we cannot
reuse any of the existing constructors to set the type to Nil.

BUG: 24335496
TEST: Tested with Python's xmlrpclib with "allow_none" option set to True
and passing a "None" value as a RPC method's parameter.

Change-Id: I6d02dbf0cb54cf1bd84015e316a23bed6b04bb55